Refactor daemon into its own package
