+++ jfr-daemon/src/main/java/com/newrelic/jfr/daemon/AttributeNames.java
@@ -1,4 +1,4 @@
-package com.newrelic.jfr;
+package com.newrelic.jfr.daemon;
 
 public class AttributeNames {
   public static final String APP_NAME = "app.name";
