+++ jfr-daemon/src/main/java/com/newrelic/jfr/daemon/EnvironmentVars.java
@@ -1,4 +1,4 @@
-package com.newrelic.jfr;
+package com.newrelic.jfr.daemon;
 
 public class EnvironmentVars {
